Move releases to GitHub Actions
Instead of continuing using Travis, which seems to be kind of
dead, I am continuing moving the release process to PyPI to GitHub
Actions.
This provides a better integration in general and seems to be more
future proof.

I've used this somewhat official guide, but I am thinking of how
to test it the best way, besides doing a production release:
https://github.com/pypa/gh-action-pypi-publish